pyproject.toml
[project]
name = "advanced-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "apache-airflow==2.10.4",
    "apache-airflow-providers-mysql==5.7.4",
    "bentoml==1.3.21",
    "catboost==1.2.7",
    "mlflow==2.20.1",
    "mysqlclient==2.2.7",
    "numpy==1.26.4",
    "pandas==2.0.3",
    "pip==25.0.1",
    "pymysql==1.1.0",
    "python-dotenv==1.0.1",
    "scikit-learn==1.2.2",
    "tqdm==4.67.1",
]

[dependency-groups]
jupyter = [
    "ipykernel>=6.29.5",
    "ipython>=9.1.0",
    "jupyter>=1.1.1",
]
